+++ include/result/coresult.h
