--- src/coresult.cpp
@@ -1,1 +1,0 @@
-#include "coresult.h"
