--- src/coresult_with_using.cpp
@@ -1,1 +1,0 @@
-#include "coresult_with_using.h"
