--- src/result.cpp
@@ -1,1 +1,0 @@
-#include "result.h"
