--- src/result_with_using.cpp
@@ -1,1 +1,0 @@
-#include "result_with_using.h"
