+++ test/source/coresult.cpp
